WebSummarizerShortner/server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.72
botocore==1.34.72
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
ffmpeg-python==0.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Testing==0.8.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
mutagen==1.47.0
openai==1.14.3
outcome==1.3.0.post0
psycopg2-binary==2.9.9
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
youtube-transcript-api==0.6.2
yt-dlp==2024.3.10
